Flexible log pipelines
Introduces a new concept called log pipeline which allows to feed log
records through a decorator, a formatter and a writer. The decorator
concept allows to add additional information to a request record. This
is useful in order to append meta information like the requested URL
or the logged in user to a request record without introducing complex
dependencies from the logging system back to the application.

Changes the package version to 1.2-SNAPSHOT due to the incompatible
nature of these changes.

Change-Id: If20aa060f5c6aea4d952f3b0052044ba77aaaa3e