Fixes log message formatting
When logging messages with three or more placeholders, the placeholders
were not substituted correctly. In fact, only the first placeholder got
substituted because the placeholder array was treated as single argument
that happened to be an array object.

So, a log statement such as

    log.info("arg1={}, arg2={}, arg3={}", "v1", "v2", "v3");

was resulting in a log message

    "arg1=[v1, v2, v3], arg2={}, arg3={}"

instead of

    "arg1=v1, arg2=v2, arg3=v3".